requirements.txt
flask
flask-restful
flask-SQLAlchemy
flask-Migrate
python-dotenv